Fix tests and make sure onComplete runs - GH-35.
onComplete wasn't being called correctly. While it was being passed to
the onTick calls appropriately it should have been fired when the job
was explicitly stopped or when the timer ran out, not to be run again.
This wasn't happening.

Additionally, the tests for onComplete were really bad. They were
expecting that the onTick fire off the onCompletes. That is...lacking to
say the least. This has now been fixed as well.

Signed-off-by: Nick Campbell <nicholas.j.campbell@gmail.com>