GH-393, GH-394 - Remove hard limit on max iters
We had a hard limit on the number of iterations a for loop would
execution when attempting to find the next date of execution. This was
arbitrary, but necessary to deter from an infinite loop that would occur
in a few cases. However, it seems better to let the execution happen up
to an arbitrary time length instead. This allows execution of larger
ranges happen, but still should protect against the infinite loop
scenario. This _may_ cause an issue if the event loop locks up, but if
that happens then the module will be hosed anyway, so don't do that. :)

Signed-off-by: Nick Campbell <nicholas.j.campbell@gmail.com>